--- src/app/public/modules/overlay/overlay-host.component.html
@@ -1,2 +1,0 @@
-<ng-template #target>
-</ng-template>
